backend/requirements.txt
Flask
Flask-MySQLdb
Flask-Bcrypt
Flask-Session
python-dotenv
Flask-CORS
gunicorn